Changed Gallery script from Jquery to Plain JS
