--- 1/index.html
@@ -1,32 +1,0 @@
-<!doctype html>
-
-<html>
-    <head>
-        <!-- <script src="http://code.jquery.com/jquery-1.11.0.min.js%22%3E"></script> -->
-        <script src="gallery.js"></script>
-        <link rel="stylesheet" href="index.css">   
-    </head>
-    <body>
-
-        <div class="GalleryContainer">
-            <ul>
-                <li class="list active" data-filter="imagebox">All</li>
-                <li class="list" data-filter="thumbnail">Thumbnails</li>
-                <li class="list" data-filter="logo">Logos</li>
-                <li class="list" data-filter="graphic">Graphics</li>
-            </ul>
-
-            <div id="viewer"></div>
-
-            <div class="imageView">
-                <div class="imageViewPanel">
-                    <div class="imageViewSpacer"></div>
-                    <div class="imageViewClose"><img class="closeicon" src="images/closeicon.svg"></div>
-                </div>
-                <div class="imageViewImage">
-                    <img>
-                </div>
-            </div>
-        </div>
-    </body>
-</html>
