+++ gallery/index.html
@@ -11,7 +11,6 @@
 
         <link rel="stylesheet" href="../style.css">
         <link rel="stylesheet" href="index.css">
-        <script src="https://code.jquery.com/jquery-1.11.0.min.js"></script>
         <script src="index.js"></script>
     </head>
 
@@ -60,7 +59,7 @@
         <div class="imageView">
             <div class="imageViewPanel">
                 <div class="imageViewSpacer"></div>
-                <div class="imageViewClose"><img class="closeicon" src="images/closeicon.svg"></div>
+                <div class="imageViewClose"><img class="closeicon" src="../images/closeicon.svg"></div>
             </div>
             <div class="imageViewImage">
                 <img>
